tools/data-ingestion/requirements.txt
# Core dependencies
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.23

# API clients
aiohttp>=3.9.0

# Utilities
tenacity>=8.2.3
ratelimit>=2.2.1
click>=8.1.7

# Development
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.12.0
mypy>=1.7.1
